cleanup mpp-utils
