+++ libraries/apollo-mpp-utils/src/wasmJsMain/kotlin/com/apollographql/apollo3/mpp/utils.wasmJs.kt
@@ -5,14 +5,3 @@ actual fun currentTimeMillis(): Long {
   return currentTimeMillis2().toLong()
 }
 
-actual fun currentTimeFormatted(): String = js("(new Date()).toISOString()")
-
-actual fun currentThreadId(): String {
-  return "wasm-js"
-}
-
-actual fun currentThreadName(): String {
-  return currentThreadId()
-}
-
-actual fun platform() = Platform.WasmJs
